Adds overflow:hidden to .box selector.
Without it, the .box would not be as tall as its content. While
that works fine for Askbot's default skin (boxes are white on a
white background), it will fail on skins in which boxes have a border
or a different background color. There's no reason not to fix this
in the default skin.