Use the user primary language when delivering badge messages
Badge awards can be triggered in the context of a different user (e.g.
when he votes on a post) and as such, we need to make sure that the
language of the notification of the user matches its settings and not
the currently active language of the user executing the triggering
action.