Fix openid association expiration issue
The existing code not handled properly the expiration time of
openid association handle. This caused a sign-in problems with
openstackid.org provider, meanwhile the launchpad provider was
not sensitive for this type of error.

Proper association issue time and lifetime variable assignment in
DjangoOpenIDStore.storeAssociation() function solves this issue
and requests a new association id when the existing one already
expired.